add collide and stick function and arena
Now, it's possible to drop element to arena.
Arrangment of the elements on arena put on Array.
We can put elements on other elements and bottom border.
Merge of elements with left or right border is impossible.
We can move elements only on arena.
Exit of elements outside from arena is impossible.